IPFS-70: lifted out offchain storage related types and functions, adjusted test classes
