+++ frame/tds-ipfs-core/src/tests.rs
@@ -85,8 +85,8 @@ fn test_cat_bytes() {
 #[test]
 fn test_ipfs_command_type_equality() {
 	ExtBuilder::default().build_and_execute_for_offchain(|| {
-		let cmd_add_bytes_one = IpfsCommand::AddBytes(vec![1,2,3,4,5], 1);
-		let cmd_add_bytes_two = IpfsCommand::AddBytes(vec![3,4,6,8,9], 0);
+		let cmd_add_bytes_one = IpfsCommand::AddBytes(1);
+		let cmd_add_bytes_two = IpfsCommand::AddBytes(0);
 		let cmd_cat_bytes = IpfsCommand::CatBytes(vec![3,4,6,8,9]);
 
 		let cmp_add_bytes = cmd_add_bytes_one.eq_type(&cmd_add_bytes_two);
